# Redis profiling experiment