--- a/libbeat/common/string.go
+++ b/libbeat/common/string.go
@@ -1,0 +1,10 @@
+package common
+
+type NetString []byte
+
+// implement encoding.TextMarshaller interface to treat []byte as raw string
+// by other encoders/serializers (e.g. JSON)
+
+func (n NetString) MarshalText() ([]byte, error) {
+	return n, nil
+}