--- a/libbeat/publisher/publish.go
+++ b/libbeat/publisher/publish.go
@@ -27,8 +27,8 @@ var debug = logp.MakeDebug("publish")
 
 // EventPublisher provides the interface for beats to publish events.
 type eventPublisher interface {
-	PublishEvent(ctx *context, event common.MapStr) bool
-	PublishEvents(ctx *context, events []common.MapStr) bool
+	PublishEvent(ctx context, event common.MapStr) bool
+	PublishEvents(ctx context, events []common.MapStr) bool
 }
 
 type context struct {