--- a/packetbeat/protos/redis/redis_parse.go
+++ b/packetbeat/protos/redis/redis_parse.go
@@ -1,8 +1,6 @@
 package redis
 
 import (
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/elastic/beats/libbeat/common"
@@ -25,10 +23,10 @@ type redisMessage struct {
 
 	IsRequest bool
 	IsError   bool
-	Message   string
-	Method    string
-	Path      string
 	Size      int
+	Message   common.NetString
+	Method    common.NetString
+	Path      common.NetString
 
 	next *redisMessage
 }
@@ -37,6 +35,12 @@ const (
 	START = iota
 	BULK_ARRAY
 	SIMPLE_MESSAGE
+)
+
+var (
+	empty    = common.NetString("")
+	emptyArr = common.NetString("[]")
+	nilStr   = common.NetString("nil")
 )
 
 // Keep sorted for future command addition
@@ -201,8 +205,40 @@ var redisCommands = map[string]struct{}{
 	"ZUNIONSTORE":      {},
 }
 
-func isRedisCommand(key string) bool {
-	_, exists := redisCommands[strings.ToUpper(key)]
+var maxCommandLen = 0
+
+const commandLenBuffer = 50
+
+func init() {
+	for k := range redisCommands {
+		l := len(k)
+		if l > maxCommandLen {
+			maxCommandLen = l
+		}
+	}
+
+	// panic normally triggered during testing to give a note about small buffer sizes
+	if maxCommandLen > commandLenBuffer {
+		panic("commandLenBuffer small")
+	}
+}
+
+func isRedisCommand(key common.NetString) bool {
+	if len(key) > maxCommandLen {
+		return false
+	}
+
+	// key to upper into pre-allocated buffer (commands use ASCII only)
+	var buf [commandLenBuffer]byte
+	upper := buf[:len(key)]
+	for i, b := range key {
+		if 'a' <= b && b <= 'z' {
+			b = b - 'a' + 'A'
+		}
+		upper[i] = b
+	}
+
+	_, exists := redisCommands[string(upper)]
 	return exists
 }
 
@@ -229,12 +265,12 @@ func (parser *parser) parse(buf *streambuf.Buffer) (bool, bool) {
 	return true, true
 }
 
-func (p *parser) dispatch(depth int, buf *streambuf.Buffer) (string, bool, bool, bool) {
+func (p *parser) dispatch(depth int, buf *streambuf.Buffer) (common.NetString, bool, bool, bool) {
 	if buf.Len() == 0 {
-		return "", false, true, false
+		return empty, false, true, false
 	}
 
-	var value string
+	var value common.NetString
 	var iserror, ok, complete bool
 	snapshot := buf.Snapshot()
 
@@ -252,7 +288,7 @@ func (p *parser) dispatch(depth int, buf *streambuf.Buffer) (string, bool, bool,
 		value, ok, complete = p.parseSimpleString(buf)
 	default:
 		debug("Unexpected message starting with %s", buf.Bytes()[0])
-		return "", false, false, false
+		return empty, false, false, false
 	}
 
 	if !ok || !complete {
@@ -261,100 +297,108 @@ func (p *parser) dispatch(depth int, buf *streambuf.Buffer) (string, bool, bool,
 	return value, iserror, ok, complete
 }
 
-func (p *parser) parseInt(buf *streambuf.Buffer) (string, bool, bool) {
-	line, err := buf.UntilCRLF()
-	if err != nil {
-		return "", true, false
+func (p *parser) parseInt(buf *streambuf.Buffer) (common.NetString, bool, bool) {
+	value, ok, complete := p.parseSimpleString(buf)
+	if ok && complete {
+		if _, err := parseInt(value); err != nil {
+			logp.Err("Failed to read integer reply: %s", err)
+		}
 	}
-
-	number := string(line[1:])
-	if _, err := strconv.ParseInt(number, 10, 64); err != nil {
-		logp.Err("Failed to read integer reply: %s", err)
-	}
-
-	return number, true, true
+	return value, ok, complete
 }
 
-func (p *parser) parseSimpleString(buf *streambuf.Buffer) (string, bool, bool) {
+func (p *parser) parseSimpleString(buf *streambuf.Buffer) (common.NetString, bool, bool) {
 	line, err := buf.UntilCRLF()
 	if err != nil {
-		return "", true, false
+		return empty, true, false
 	}
 
-	return string(line[1:]), true, true
+	return common.NetString(line[1:]), true, true
 }
 
-func (p *parser) parseString(buf *streambuf.Buffer) (string, bool, bool) {
+func (p *parser) parseString(buf *streambuf.Buffer) (common.NetString, bool, bool) {
 	line, err := buf.UntilCRLF()
 	if err != nil {
-		return "", true, false
+		return empty, true, false
 	}
 
 	if len(line) == 3 && line[1] == '-' && line[2] == '1' {
-		return "nil", true, true
+		return nilStr, true, true
 	}
 
-	length, err := strconv.ParseInt(string(line[1:]), 10, 64)
+	length, err := parseInt(line[1:])
 	if err != nil {
 		logp.Err("Failed to read bulk message: %s", err)
-		return "", false, false
+		return empty, false, false
 	}
 
 	content, err := buf.CollectWithSuffix(int(length), []byte("\r\n"))
 	if err != nil {
 		if err != streambuf.ErrNoMoreBytes {
-			return "", false, false
+			return common.NetString{}, false, false
 		}
-		return "", true, false
+		return common.NetString{}, true, false
 	}
 
-	return string(content), true, true
+	return common.NetString(content), true, true
 }
 
-func (p *parser) parseArray(depth int, buf *streambuf.Buffer) (string, bool, bool, bool) {
+func (p *parser) parseArray(depth int, buf *streambuf.Buffer) (common.NetString, bool, bool, bool) {
 	line, err := buf.UntilCRLF()
 	if err != nil {
 		debug("End of line not found, waiting for more data")
-		return "", false, false, false
+		return empty, false, false, false
 	}
 	debug("line %s: %d", line, buf.BufferConsumed())
 
 	if len(line) == 3 && line[1] == '-' && line[2] == '1' {
-		return "nil", false, true, true
+		return nilStr, false, true, true
 	}
 
 	if len(line) == 2 && line[1] == '0' {
-		return "[]", false, true, true
+		return emptyArr, false, true, true
 	}
 
-	count, err := strconv.ParseInt(string(line[1:]), 10, 64)
+	count, err := parseInt(line[1:])
 	if err != nil {
 		logp.Err("Failed to read number of bulk messages: %s", err)
-		return "", false, false, false
+		return empty, false, false, false
 	}
 	if count < 0 {
-		return "nil", false, true, true
+		return nilStr, false, true, true
 	} else if count == 0 {
 		// should not happen, but handle just in case ParseInt did return 0
-		return "[]", false, true, true
+		return emptyArr, false, true, true
 	}
 
 	// invariant: count > 0
-	content := make([]string, 0, count)
+
+	// try to allocate content array right on stack
+	var content [][]byte
+	const arrayBufferSize = 32
+	if int(count) <= arrayBufferSize {
+		var arrayBuffer [arrayBufferSize][]byte
+		content = arrayBuffer[:0]
+	} else {
+		content = make([][]byte, 0, count)
+	}
+
+	contentLen := 0
 	// read sub elements
 
 	iserror := false
 	for i := 0; i < int(count); i++ {
-		var value string
+		var value common.NetString
 		var ok, complete bool
 
 		value, iserror, ok, complete := p.dispatch(depth+1, buf)
 		if !ok || !complete {
 			debug("Array incomplete")
-			return "", iserror, ok, complete
+			return empty, iserror, ok, complete
 		}
 
-		content = append(content, value)
+		content = append(content, []byte(value))
+		contentLen += len(value)
 	}
 
 	// handle top-level request command
@@ -365,11 +409,38 @@ func (p *parser) parseArray(depth int, buf *streambuf.Buffer) (string, bool, boo
 			p.message.Path = content[1]
 		}
 
-		value := strings.Join(content, " ")
+		var value common.NetString
+		if contentLen > 1 {
+			tmp := make([]byte, contentLen+(len(content)-1)*1)
+			join(tmp, content, []byte(" "))
+			value = common.NetString(tmp)
+		} else {
+			value = common.NetString(content[0])
+		}
 		return value, iserror, true, true
 	}
 
-	// return redis array
-	value := "[" + strings.Join(content, ", ") + "]"
+	// return redis array: [a, b, c]
+	tmp := make([]byte, 2+contentLen+(len(content)-1)*2)
+	tmp[0] = '['
+	join(tmp[1:], content, []byte(", "))
+	tmp[len(tmp)-1] = ']'
+	value := common.NetString(tmp)
 	return value, iserror, true, true
+}
+
+func parseInt(line []byte) (int64, error) {
+	buf := streambuf.NewFixed(line)
+	return buf.AsciiInt(false)
+	// TODO: is it an error if 'buf.Len() != 0 {}' ?
+}
+
+func join(to []byte, content [][]byte, sep []byte) {
+	if len(content) > 0 {
+		off := copy(to, content[0])
+		for i := 1; i < len(content); i++ {
+			off += copy(to[off:], sep)
+			off += copy(to[off:], content[i])
+		}
+	}
 }