… libbeat/publisher/common_test.go
… packetbeat/CHANGELOG.md
… packetbeat/protos/redis/redis_test.go