Did play with golang pprof tool today + redis pcap files. This PR hugely reduces number of objects allocated by redis-module itself. When replaying redis traffic only object allocations are mostly dominated by the json encoder (the output worker directly calls console outputer doing json encoding):

    (pprof) top20 -cum
    2809008 of 6861983 total (40.94%)
    Dropped 48 nodes (cum <= 34309)
    Showing top 20 nodes out of 61 (cum >= 1385433)
        flat  flat%   sum%        cum   cum%
            0     0%     0%    6860618   100%  runtime.goexit
            0     0%     0%    3767472 54.90%  github.com/elastic/beats/libbeat/publisher.(*messageWorker).run
            0     0%     0%    2834909 41.31%  github.com/elastic/beats/libbeat/publisher.(*outputWorker).onBulk
            0     0%     0%    2834909 41.31%  github.com/elastic/beats/libbeat/publisher.(*outputWorker).onMessage
        229379  3.34%  3.34%    2834909 41.31%  github.com/elastic/beats/libbeat/publisher.(*outputWorker).sendBulk
        56608  0.82%  4.17%    2605530 37.97%  encoding/json.Marshal
            0     0%  4.17%    2605530 37.97%  github.com/elastic/beats/libbeat/outputs.(*bulkOutputAdapter).BulkPublish
            0     0%  4.17%    2605530 37.97%  github.com/elastic/beats/libbeat/outputs/console.(*console).PublishEvent
            0     0%  4.17%    2578716 37.58%  encoding/json.(*encodeState).marshal
            0     0%  4.17%    2578716 37.58%  encoding/json.(*encodeState).reflectValue
            0     0%  4.17%    2578716 37.58%  encoding/json.(*mapEncoder).(encoding/json.encode)-fm
        81922  1.19%  5.36%    2578716 37.58%  encoding/json.(*mapEncoder).encode
            0     0%  5.36%    2220027 32.35%  github.com/elastic/beats/packetbeat/beat.(*Packetbeat).Run.func1
        196612  2.87%  8.23%    2220027 32.35%  github.com/elastic/beats/packetbeat/sniffer.(*SnifferSetup).Run
        71010  1.03%  9.26%    1849392 26.95%  github.com/elastic/beats/packetbeat/decoder.(*DecoderStruct).DecodePacketData
        377929  5.51% 14.77%    1774285 25.86%  github.com/elastic/beats/packetbeat/protos/tcp.(*Tcp).Process
    1458197 21.25% 36.02%    1458197 21.25%  reflect.unsafe_New
            0     0% 36.02%    1396356 20.35%  github.com/elastic/beats/packetbeat/protos/redis.(*Redis).Parse
            0     0% 36.02%    1396356 20.35%  github.com/elastic/beats/packetbeat/protos/tcp.(*TcpStream).addPacket
        337351  4.92% 40.94%    1385433 20.19%  github.com/elastic/beats/packetbeat/protos/redis.(*Redis).doParse
  
TODO: some benchmarking on redis module to check if allocations really make a big difference.

changes
=======

- added commandline flag '-httpprof' to start HTTP pprof endpoint for doing
  some live profiling of beat services (heap profiling, cpu profiling)

- add streambuf.AppendWithCapLimits which guarantees newly allocated
  buffers have at least configured capacity in case the buffer
  must grow when appending new content (excert more control on buffer sizing)
  => not used, as number of objects allocated was minor in contrast to
     other object types. Plus bad policy can increase overal memory usage
     dramatically due to pointers back into original slice

- add common.NetString basically being []byte. Difference to []byte is, this
  type implements the encoding.TextMarhsaller interface. NetString objects
  will be handled by JSON encoder like normal strings, thusly no allocation
  and copy is required when extracting strings from parsed messages.
  NetString being []byte will directly point into original message buffer.

  NetString also provides UnsafeString to cast a NetString into a string
  without allocation + copy

  => Using NetString over string reduce allocations + copy operations +
     less object to be GCed

- remove some unwanted allocation in async publisher client

- redis 'optimizations':
  - replace all string types with common.NetString to reduce allocations
  - pre-allocate command buffer holding command + 4 arguments by default
  - remove some more allocations + copies by 'low-level code'
  - remove transaction object type (not required anymore due to correlation
    supporting piping), but generate event type directly from correlated
    messages => reduces allocation once more + removes computation of unused
    fields