@andrewkroh 2015-12-04
I like the `-httpprof` flag. I am always modifying the source to drop in the HTTP server to enable pprof.