@urso 2015-12-04
added some very simple/small benchmarks for redis parser only (does not take the removed allocations for removal of redundant transaction object into account):

old parser (using string):

    BenchmarkParserNoArgsResult-4 	 1000000	      1414 ns/op	     464 B/op	       9 allocs/op
    BenchmarkParserArrayRequest-4 	 1000000	      2108 ns/op	     592 B/op	      15 allocs/op
    BenchmarkParserArrayResponse-4	 1000000	      2420 ns/op	     688 B/op	      17 allocs/op
    BenchmarkParserArrayNested-4  	 1000000	      3223 ns/op	     896 B/op	      26 allocs/op
    BenchmarkParserArray2Passes-4 	  500000	      3390 ns/op	     832 B/op	      24 allocs/op

new parser (using NetString):

    BenchmarkParserNoArgsResult-4 	 1000000	      1338 ns/op	     448 B/op	       6 allocs/op
    BenchmarkParserArrayRequest-4 	 1000000	      1551 ns/op	     448 B/op	       6 allocs/op
    BenchmarkParserArrayResponse-4	 1000000	      1598 ns/op	     464 B/op	       6 allocs/op
    BenchmarkParserArrayNested-4  	 1000000	      2263 ns/op	     640 B/op	      13 allocs/op
    BenchmarkParserArray2Passes-4 	 1000000	      2302 ns/op	     576 B/op	      10 allocs/op

Allocations and runtime are reduced (some allocations are due to test setup).