@tsg 2015-12-07
LGTM! Do you want to cleanup history before merging?