@tsg 2015-12-07
Also a line in the Changelog about the Redis parser being faster would be good.