@urso 2015-12-08
cleaned up history + guarded all debug statements.

New benchmark results:

    BenchmarkParserNoArgsResult-4 	 1000000	      1069 ns/op	     368 B/op	       3 allocs/op
    BenchmarkParserArrayRequest-4 	 1000000	      1292 ns/op	     368 B/op	       3 allocs/op
    BenchmarkParserArrayResponse-4	 1000000	      1331 ns/op	     384 B/op	       3 allocs/op
    BenchmarkParserArrayNested-4  	 1000000	      1457 ns/op	     400 B/op	       4 allocs/op
    BenchmarkParserArray2Passes-4 	 1000000	      1804 ns/op	     416 B/op	       4 allocs/op