@ruflin 2015-12-08
@urso As @travis-ci seems to be broken at I suggest to rely on our Jenkins build.